Robot moves

+++ GraphicObject.cpp
@@ -12,7 +12,7 @@ GraphicObject::GraphicObject()
     orientationAngle = 0;
 }
 
-void GraphicObject::setLocation(int x, int y)
+void GraphicObject::setLocation(double x, double y)
 {
     location[0] = x;
     location[1] = y;
